+++ monarchy/src/main/java/com/zhuinden/monarchy/ManagedLiveResults.java
@@ -19,6 +19,57 @@ import io.realm.RealmResults;
 
 class ManagedLiveResults<T extends RealmModel>
         extends MutableLiveData<Monarchy.ManagedChangeSet<T>> {
+    private static class EmptyChangeSet implements OrderedCollectionChangeSet {
+        private static final int[] NO_INDEX_CHANGES = new int[0];
+        private static final Range[] NO_RANGE_CHANGES = new Range[0];
+        
+        @Override
+        public State getState() {
+            return State.INITIAL;
+        }
+
+        @Override
+        public int[] getDeletions() {
+            return NO_INDEX_CHANGES;
+        }
+
+        @Override
+        public int[] getInsertions() {
+            return NO_INDEX_CHANGES;
+        }
+
+        @Override
+        public int[] getChanges() {
+            return NO_INDEX_CHANGES;
+        }
+
+        @Override
+        public Range[] getDeletionRanges() {
+            return NO_RANGE_CHANGES;
+        }
+
+        @Override
+        public Range[] getInsertionRanges() {
+            return NO_RANGE_CHANGES;
+        }
+
+        @Override
+        public Range[] getChangeRanges() {
+            return NO_RANGE_CHANGES;
+        }
+
+        @Nullable
+        @Override
+        public Throwable getError() {
+            return null;
+        }
+
+        @Override
+        public boolean isCompleteResult() {
+            return true;
+        }
+    }
+    
     private final Monarchy monarchy;
     private final Monarchy.Query<T> query;
     private final boolean asAsync;
@@ -47,11 +98,15 @@ class ManagedLiveResults<T extends RealmModel>
         realm = Realm.getInstance(realmConfiguration);
         RealmQuery<T> realmQuery = query.createQuery(realm);
         if(asAsync) {
-            realmResults = realmQuery.findAllAsync(); // sort/distinct should be done with new queries
+            realmResults = realmQuery.findAllAsync(); // sort/distinct should be done with new queries, 5.0+
         } else {
-            realmResults = realmQuery.findAll(); // sort/distinct should be done with new queries
+            realmResults = realmQuery.findAll(); // sort/distinct should be done with new queries, 5.0+
         }
         realmResults.addChangeListener(realmChangeListener);
+
+        if(!asAsync) {
+            setValue(new Monarchy.ManagedChangeSet<T>(realmResults, new EmptyChangeSet()));
+        }
     }
 
     @Override